+++ packages/eip712/src/payload/types/index.ts
@@ -1,0 +1,1 @@
+export { default } from './types.js'
